Unify admin behavior

+++ core/src/main/java/com/scalar/db/transaction/consensuscommit/ConsensusCommitAdmin.java
@@ -54,18 +54,30 @@ public class ConsensusCommitAdmin implements DistributedTransactionAdmin {
 
   @Override
   public void createCoordinatorTables(Map<String, String> options) throws ExecutionException {
+    if (coordinatorTablesExist()) {
+      throw new IllegalArgumentException("coordinator tables already exist");
+    }
+
     admin.createNamespace(coordinatorNamespace, options);
     admin.createTable(coordinatorNamespace, Coordinator.TABLE, Coordinator.TABLE_METADATA, options);
   }
 
   @Override
   public void dropCoordinatorTables() throws ExecutionException {
+    if (!coordinatorTablesExist()) {
+      throw new IllegalArgumentException("coordinator tables do not exist");
+    }
+
     admin.dropTable(coordinatorNamespace, Coordinator.TABLE);
     admin.dropNamespace(coordinatorNamespace);
   }
 
   @Override
   public void truncateCoordinatorTables() throws ExecutionException {
+    if (!coordinatorTablesExist()) {
+      throw new IllegalArgumentException("coordinator tables do not exist");
+    }
+
     admin.truncateTable(coordinatorNamespace, Coordinator.TABLE);
   }
 
@@ -165,6 +177,9 @@ public class ConsensusCommitAdmin implements DistributedTransactionAdmin {
       String namespace, String table, String columnName, DataType columnType)
       throws ExecutionException {
     TableMetadata tableMetadata = getTableMetadata(namespace, table);
+    if (tableMetadata == null) {
+      throw new IllegalArgumentException("Table does not exist: " + namespace + "." + table);
+    }
     String beforeColumnName = getBeforeImageColumnName(columnName, tableMetadata);
 
     admin.addNewColumnToTable(namespace, table, columnName, columnType);
